Create PostgreSQL service
The database pod is running, but in order to connect to another pod we need to add the service.

Kubernetes pods are transient - they can be killed, restarted, or created dynamically. Therefore we don't want to try to connect to pods directly, but rather create a Kubernetes Service. Services keep track of pods and direct traffic to the right place.